factor out request_factory
